More secure AUTH
This commit introduces more secure practices, the JWTSecret key is now
set using the environment variables and rather than using a single
stateless JWT token, we are now using two. A refresh token and access
token.

The access token is short lived for 15 minutes and is required for using
the end points, the refresh token is required to generate the access
token each time and is valid for 7 days. This token is stored in the
database.

Future updates should explore the possibility of a token blacklist or
revokation mechanism.